task(feature/auth-2fa-fixes): Fix logn / 2fa

+++ src/app/pages/auth/organization-login/organization-login.component.ts
@@ -86,7 +86,7 @@ export class OrganizationLoginComponent implements OnInit {
           localStorage.setItem('authData', JSON.stringify(data));
           this.notificationService.publishMessages('success', 'Login Successful');
             this.router.navigateByUrl('/organization/dashboard');
-        }  else {
+        }  else if (this.loggedInUser.UserType !== 'Organization') {
           this.notificationService.publishMessages('error', 'Invalid login credential');
           localStorage.clear()
 
@@ -94,7 +94,7 @@ export class OrganizationLoginComponent implements OnInit {
         }
       } else {
         this.show2FAOTP = true
-        this.timer(10)
+        this.timer(1)
      
       }
     })
@@ -174,7 +174,7 @@ export class OrganizationLoginComponent implements OnInit {
     this.actions$.pipe(ofType(confirm2FActionSuccess)).subscribe((res: any) => {
       if (res.message.hasErrors === false) {
         this.notificationService.publishMessages('success', res.message.description);
-        this.timer(10)
+        this.timer(1)
        
 
       }
